Remove all types, create one type for commands called 'COMMAND'
Having actions based on type was very inefficient. Terminal commands consist of a commmand, a set of flags, and a directory path. All terminal commands fall into one of two categories those that change the directory tree or the file path (cd, rm, mkdir etc.), and those that don't (ls, cat, pwd etc). It miss much cleaner to make this distinction and handle all the commands that change the state (like cd, rm, mkdir) in the reducer and handle all the other in the render statement in this file. This involved a major rewrite to all the code.